feat(ci): fix to mocha cli, fix coveralls
